tests/conftest.py
import pytest
import os
from os import getenv

from flask import Flask
from flask_sqlalchemy import SQLAlchemy

from dotenv import load_dotenv

BASEDIR = os.path.abspath(os.path.dirname(__file__))
load_dotenv(os.path.join(BASEDIR, '.env'))

app = Flask(__name__)

app.config["SQLALCHEMY_DATABASE_URI"] = getenv("DATABASE_URL")
app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False

db = SQLAlchemy(app)

# Create fixture from this later when you have sorted out the Postgres problems