requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==21.12b0
bleach==4.1.0
cffi==1.15.0
click==8.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.3
executing==0.8.2
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
pyzmq==22.3.0
Send2Trash==1.8.0
six==1.16.0
SQLAlchemy==1.3.23
stack-data==0.1.4
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
zipp==3.7.0
